--- manager/controllers/app/config_evaluator_api.go
@@ -1,4 +1,0 @@
-// Copyright 2020 IBM Corp.
-// SPDX-License-Identifier: Apache-2.0
-
-package app
