+++ manager/controllers/app/dataset/data_info.go
@@ -1,7 +1,7 @@
 // Copyright 2020 IBM Corp.
 // SPDX-License-Identifier: Apache-2.0
 
-package asset_details
+package dataset
 
 import (
 	"fybrik.io/fybrik/pkg/serde"
