+++ manager/controllers/app/evaluator/evaluator_interface.go
@@ -1,7 +1,7 @@
 // Copyright 2020 IBM Corp.
 // SPDX-License-Identifier: Apache-2.0
 
-package config_evaluator
+package evaluator
 
 import corev1 "k8s.io/api/core/v1"
